BLD: introduce an `as_subproject` build option
This makes it a lot easier to use one or more shared libraries
built within a Meson subproject in the main project. Without
this, meson-python will choke on some of the install targets
that it doesn't know what to do with in a wheel.

Also convert some of the utility libraries to static libraries
- they're private so no need for them to be dynamic.

Finally, remove `both_libraries` (we need only one type typically,
although we may have to handle some build options better) and add
more dependencies that can be used from a main project.